refactor: 가위바위보와 묵찌빠를 별개의 구조체로 분리
